# docs/swagger/atlas-api.yaml
...
openapi: 3.0.3
info:
  title: ScaleMargin Dispatcher API — Atlas data plane
  version: "1.0.0"
  description: |
    The HTTP API the **ScaleMargin Atlas** platform calls to render a dispatcher
    dashboard, author its variables, and read campaign reporting and logs.

    **The contract in one line:** the base URL changes per client, the endpoints
    never do. Atlas hard-codes the paths and varies only the host.

    ## Two routers

    | Tree | Audience | Auth | Reachability |
    | --- | --- | --- | --- |
    | `/api/v1/data-plane/*` | Atlas | API key | must be reachable from ScaleMargin |
    | `/api/v1/internal/*` | the client's own infrastructure | none | should **not** be internet-exposed |

    ## Authentication

    One environment variable on the dispatcher (`DISPATCHER_ATLAS_KEY`, 32+
    chars, e.g. `openssl rand -base64 32`). No database record, no UI — the
    deployment owns the credential.

    * Send it as `Authorization: Bearer <key>`. **Never in a query string** —
      it lands in access logs, proxy logs and browser history.
    * Comparison is constant-time over SHA-256 digests; neither the key's length
      nor its prefix leaks through response timing.
    * **Rotation/revocation = change the value and restart.** No runtime kill switch.
    * **Unset means disabled:** every data-plane route returns `503 unavailable`
      while `DISPATCHER_ATLAS_KEY` is absent — including requests presenting a
      key. It fails closed and never falls open.
...

...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/data-plane/variables:
    get:
      operationId: listVariables
      summary: List variables (the catalog)
      description: |
        Names, definitions and samples; never a resolved value. Sorted by name
        ascending. Filters narrow the set, then the page is cut from what
        survives, so `meta.total` describes the filtered result — never the whole
        table. A page past the end is a `200` with an empty array and honest
        counts; walk back rather than handling a 404. `total_pages` floors at 1.
      tags: [Variables]
      parameters:
        - $ref: "#/components/parameters/sourceFilter"
        - $ref: "#/components/parameters/enabledFilter"
        - $ref: "#/components/parameters/nameSearch"
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/limitParam"
      responses:
        "200":
          description: A filtered page of variables
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VariableListResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/Unavailable" }
        "500": { $ref: "#/components/responses/Internal" }
    post:
      operationId: createVariable
      summary: Create a variable
      description: |
        Nothing is written when validation fails (`400 invalid_request` with a
        details entry pointing at the offending field). Validation highlights:

        * `name`: no spaces, dots or dashes; must start with a letter or underscore.
        * `query.sql`: a **single** SELECT/WITH statement — a semicolon is
          rejected, so nothing can be stacked onto a read.
        * `computed.expr`: checked against the real evaluator, so anything
          accepted here also resolves at send time.
        * `api.url`: must start with http(s); {{tokens}} are interpolated
          URL-encoded at send time.
        * A masked header value (`••••••••`) is rejected on create — there is no
          stored secret to preserve yet. Send the real value.
        * `source` must appear in `lookup.supported_variable_sources` from
          `/state`. In network and mock mode there is no SQL connection, so
          `query` is refused with `400 invalid_request` on
          `details[].path = "definition.source"`. Read that list rather than
          hard-coding the rule.

        `sample` is ignored for field/computed/constant — the dispatcher renders
        those against a fictional sample record itself. For query/api it is
        stored as sent (a save must never run the client's SQL or call their
        internal service); send whatever your live test returned.

        On success the placeholder snapshot refreshes: applies to the **next
        dispatch**, no restart.
      tags: [Variables]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVariableRequest"
            example:
              name: available_credit
              definition:
                source: query
                sql: "SELECT limit_amount - drawn_amount FROM credit_accounts WHERE user_ref = {{user_id}}"
              fallback: your current limit
              sample: "48,250.00"
              enabled: true
      responses:
        "201":
          description: Created — echoes the stored variable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SingleVariableResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/Unavailable" }
        "500": { $ref: "#/components/responses/Internal" }
  /api/v1/data-plane/variables/{name}:
    parameters:
      - $ref: "#/components/parameters/variableName"
    get:
      operationId: getVariable
      summary: Get one variable by placeholder name
      tags: [Variables]
      responses:
        "200":
          description: The variable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SingleVariableResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/Unavailable" }
        "500": { $ref: "#/components/responses/Internal" }
    patch:
      operationId: updateVariable
      summary: Partially update a variable
      description: |
        Partial by field, atomic by definition. Every key is optional but at
        least one is required (an empty body is 400).

        * Omit definition → recipe untouched. Send it → replaced **whole** (a
          partial definition would let a caller turn a query into a field while
          leaving the SQL behind). Switching source clears the previous source's
          columns.
        * Rename via `name`; renaming onto an existing name is 409 conflict.
        * **Masked headers mean "keep the stored secret":** sending
          `"••••••••"` back preserves the stored value, so an editor that can
          never read a bearer token can still round-trip its definition. A
          header the caller omits is dropped.
        * `enabled: false` disables resolution while keeping the definition.
        * A replaced definition is checked against
          `lookup.supported_variable_sources` the same way a create is, so
          switching an existing variable to `query` in network mode is refused.
          Variables already stored as `query` are left alone — switching lookup
          mode never deletes them; they sit inactive on their fallback until a
          SQL connection returns.
      tags: [Variables]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVariableRequest"
            examples:
              disableOnly:
                summary: Toggle status without touching the definition
                value:
                  enabled: false
              renameAndReplace:
                summary: Rename and replace the definition whole
                value:
                  name: credit_available
                  definition:
                    source: constant
                    value: "—"
      responses:
        "200":
          description: Updated variable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SingleVariableResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/Unavailable" }
        "500": { $ref: "#/components/responses/Internal" }
    delete:
      operationId: deleteVariable
      summary: Delete a variable
      description: >
        Immediate and not reversible. Templates still referencing the placeholder
        render it as an empty string on the next dispatch. 404 if already gone.
      tags: [Variables]
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteVariableResponse"
              example:
                deleted: true
                name: available_credit
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/Unavailable" }
        "500": { $ref: "#/components/responses/Internal" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        The dispatcher's DISPATCHER_ATLAS_KEY (32+ chars). Header only — never
        in a query string. Unset on the dispatcher, every route answers 503
        unavailable regardless of the key presented.
  parameters:
    pageParam:
      name: page
      in: query
      required: false
      description: 1-based page number; filters apply before the page is cut
      schema: { type: integer, minimum: 1, default: 1 }
    limitParam:
      name: limit
      in: query
      required: false
      description: Rows per page — capped so one request cannot serialize the whole table
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    sourceFilter:
      name: source
      in: query
      required: false
      description: Filter by definition source type
      schema: { $ref: "#/components/schemas/VariableSource" }
    enabledFilter:
      name: enabled
      in: query
      required: false
      description: Filter by status; omitted returns both
      schema: { type: boolean }
    nameSearch:
      name: q
      in: query
      required: false
      description: Case-insensitive substring match on the placeholder name
      schema: { type: string, maxLength: 191 }
    variableName:
      name: name
      in: path
      required: true
      description: The placeholder name — what the template writes between {{ }}
      schema:
        type: string
        minLength: 1
        maxLength: 191
        pattern: "^[a-zA-Z_][a-zA-Z0-9_]*$"
...
  responses:
    BadRequest:
      description: Payload failed validation — see details[].path
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example:
            error: invalid_request
            message: Request failed validation
            details:
              - path: definition.sql
                message: SQL must be a SELECT or WITH statement
    Unauthorized:
      description: Missing, malformed or wrong key — same body for every reason
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example:
            error: unauthorized
            message: Valid API key required
    NotFound:
      description: Unknown path, or a named resource that does not exist
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example:
            error: not_found
            message: Variable "nope" does not exist
    Conflict:
      description: The write would duplicate a name that already exists
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example:
            error: conflict
            message: Variable "available_credit" already exists
    RateLimited:
      description: More than 60 requests per minute for one key — back off
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example:
            error: rate_limited
            message: Rate limit is 60 requests per minute
    Unavailable:
      description: >
        Key unset (integration not enabled), no state database attached, or the
        state database unreachable where degrading was impossible
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example:
            error: unavailable
            message: "Atlas API is not configured on this dispatcher (DISPATCHER_ATLAS_KEY is unset)"
    Internal:
      description: Unexpected — should not occur; report it
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example:
            error: internal
            message: Unexpected error
  schemas:
...
    ApiErrorCode:
      type: string
      description: Stable machine-readable code — branch on this, never on prose
      enum:
        - invalid_request   # 400 payload failed validation
        - unauthorized      # 401 missing/malformed/wrong key
        - not_found         # 404 unknown path or resource
        - conflict          # 409 duplicate name
        - rate_limited      # 429 over 60 req/min/key
        - unavailable       # 503 key unset / no usable state DB
        - internal          # 500 unexpected
    ErrorDetail:
      type: object
      required: [path, message]
      properties:
        path:
          type: string
          description: Dotted field path, e.g. definition.api.headers.Authorization
        message:
          type: string
    ErrorEnvelope:
      type: object
      required: [error, message]
      properties:
        error: { $ref: "#/components/schemas/ApiErrorCode" }
        message:
          type: string
        details:
          type: array
          description: Present on invalid_request only — one entry per failing field
          items: { $ref: "#/components/schemas/ErrorDetail" }
    PageMeta:
      type: object
      required:
        - page
        - limit
        - total
        - total_pages
        - from
        - to
        - has_previous_page
        - has_next_page
      properties:
        page:
          type: integer
          description: Echoes the requested page, even when past the end
        limit:
          type: integer
          description: Echoes the requested page size
        total:
          type: integer
          description: Rows matching the FILTERS, before the page is cut
        total_pages:
          type: integer
          description: ceil(total/limit), floored at 1 — an empty result is "page 1 of 1"
        from:
          type: integer
          nullable: true
          description: 1-based index of the first row in this response; null when empty
        to:
          type: integer
          nullable: true
          description: 1-based index of the last row; null when empty
        has_previous_page: { type: boolean }
        has_next_page:
          type: boolean
          description: A page past the end is 200 with an empty array, never a 404
...
    VariableSource:
      type: string
      enum: [field, computed, constant, query, api]
...
    VariableDefinition:
      description: Discriminated union — an invalid source fails before branch rules run
      oneOf:
        - $ref: "#/components/schemas/FieldDefinition"
        - $ref: "#/components/schemas/ComputedDefinition"
        - $ref: "#/components/schemas/ConstantDefinition"
        - $ref: "#/components/schemas/QueryDefinition"
        - $ref: "#/components/schemas/ApiDefinition"
      discriminator:
        propertyName: source
        mapping:
          field: "#/components/schemas/FieldDefinition"
          computed: "#/components/schemas/ComputedDefinition"
          constant: "#/components/schemas/ConstantDefinition"
          query: "#/components/schemas/QueryDefinition"
          api: "#/components/schemas/ApiDefinition"
...
    Variable:
      type: object
      required:
        - name
        - source
        - definition
        - fallback
        - sample
        - enabled
        - created_at
        - updated_at
        - updated_by
      properties:
        name:
          type: string
          maxLength: 191
          description: The placeholder as written in a template — {{available_credit}}. Unique.
        source:
          $ref: "#/components/schemas/VariableSource"
        definition:
          $ref: "#/components/schemas/VariableDefinition"
        placeholders:
          type: array
          items: { type: string }
          description: >
            Every token a template can use for this variable — always a single
            value. For an api variable: `name` only when `json_path` is set,
            plus `name.<path>` for each `response_schema` path.
        system:
          type: boolean
          description: Defined in dispatcher code — never editable or deletable; no timestamps
        description:
          type: string
          nullable: true
          description: One line, system variables only
        fallback:
          type: string
          nullable: true
          description: Used when the source yields nothing or fails — with one, a variable can never break a send
        sample:
          type: string
          nullable: true
          description: Preview against the fictional sample record (usr_1024 / Ada Lovelace / Acme Corp). Never a real customer.
        enabled:
          type: boolean
          description: A disabled variable keeps its definition but stops resolving
        created_at: { type: string, format: date-time, nullable: true, description: null for system variables }
        updated_at: { type: string, format: date-time, nullable: true, description: null for system variables }
        updated_by:
          type: string
          nullable: true
          description: '"atlas:" plus the first 8 characters of the key that wrote it; "system" for system variables'
    CreateVariableRequest:
      type: object
      required: [name, definition]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 191
          pattern: "^[a-zA-Z_][a-zA-Z0-9_]*$"
        definition: { $ref: "#/components/schemas/VariableDefinition" }
        fallback: { type: string, nullable: true, maxLength: 4000 }
        sample:
          type: string
          nullable: true
          maxLength: 4000
          description: Ignored for field/computed/constant (rendered server-side); stored as sent for query/api
        enabled:
          type: boolean
          default: true
    UpdateVariableRequest:
      type: object
      minProperties: 1
      description: Every field optional, at least one required; definition moves whole
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 191
          pattern: "^[a-zA-Z_][a-zA-Z0-9_]*$"
        definition: { $ref: "#/components/schemas/VariableDefinition" }
        fallback: { type: string, nullable: true, maxLength: 4000 }
        sample: { type: string, nullable: true, maxLength: 4000 }
        enabled: { type: boolean }
    SingleVariableResponse:
      type: object
      required: [variable]
      properties:
        variable: { $ref: "#/components/schemas/Variable" }
    VariableListResponse:
      type: object
      required: [generated_at, meta, variables]
      properties:
        generated_at: { type: string, format: date-time }
        meta: { $ref: "#/components/schemas/PageMeta" }
        variables:
          type: array
          items: { $ref: "#/components/schemas/Variable" }
    DeleteVariableResponse:
      type: object
      required: [deleted, name]
      properties:
        deleted:
          type: boolean
          enum: [true]
        name: { type: string }
...